pyproject.toml
[tool.poetry]
name = "repl_python3_GABI2021initiationpython"
version = "0.1.0"
description = ""
authors = ["mariabernard"]

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.78"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
